consolidate tests, and remove some silly nil checks
