+++ loglater_test.go
@@ -3,53 +3,37 @@ package loglater
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"io"
 	"log/slog"
-	"sort"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
 	"github.com/robbyt/go-loglater/storage"
 )
 
-// Compile-time interface check
+// Interface compliance
 var _ slog.Handler = (*LogCollector)(nil)
 
 func TestLogCollectorImplementsSlogHandler(t *testing.T) {
 	var handler slog.Handler = NewLogCollector(nil)
-
 	_ = handler.Enabled(context.Background(), slog.LevelInfo)
 	_ = handler.WithAttrs([]slog.Attr{slog.String("key", "value")})
 	_ = handler.WithGroup("groupname")
 	_ = handler.Handle(context.Background(), slog.Record{})
 }
 
-func TestLogCollector(t *testing.T) {
-	t.Run("CollectsAndPlaysLogs", func(t *testing.T) {
-		// Create a buffer for capturing output
+// Core functionality
+func TestBasicLogging(t *testing.T) {
+	t.Run("CollectAndReplay", func(t *testing.T) {
 		var buf bytes.Buffer
-
-		// Create a JSON handler outputting to the buffer
-		jsonHandler := slog.NewJSONHandler(&buf, nil)
-
-		// Create collector with no output handler (nil)
 		collector := NewLogCollector(nil)
-
-		// Create logger that uses our collector
 		logger := slog.New(collector)
 
-		// Log some events with a fixed time
 		logger.WithGroup("app").Info("Starting up", "version", "1.0")
 		logger.Error("Something failed", "error", "test error")
 
-		// Verify nothing was written to the buffer yet
-		if buf.Len() > 0 {
-			t.Errorf("expected empty buffer before replay, got %q", buf.String())
-		}
-
-		// Get the logs and verify they were collected correctly
 		logs := collector.GetLogs()
 		if len(logs) != 2 {
 			t.Fatalf("expected 2 logs, got %d", len(logs))
@@ -63,172 +47,47 @@ func TestLogCollector(t *testing.T) {
 			t.Errorf("expected Error level, got %v", logs[1].Level)
 		}
 
-		// Now play the logs to the JSON handler
+		jsonHandler := slog.NewJSONHandler(&buf, nil)
 		if err := collector.PlayLogs(jsonHandler); err != nil {
 			t.Fatalf("unexpected error playing logs: %v", err)
 		}
 
-		// Verify that output was written to the buffer
 		if buf.Len() == 0 {
 			t.Error("expected non-empty buffer after replay")
 		}
-
-		// Parse the output into JSON
-		output := buf.String()
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		if len(lines) != 2 {
-			t.Fatalf("expected 2 lines of JSON output, got %d", len(lines))
-		}
-
-		// Test the second log entry (error message)
-		var logEntry map[string]interface{}
-		if err := json.Unmarshal([]byte(lines[1]), &logEntry); err != nil {
-			t.Fatalf("invalid JSON: %v", err)
-		}
-
-		if msg, ok := logEntry["msg"].(string); !ok || msg != "Something failed" {
-			t.Errorf("expected error message 'Something failed', got %v", logEntry["msg"])
-		}
-
-		if level, ok := logEntry["level"].(string); !ok || level != "ERROR" {
-			t.Errorf("expected level 'ERROR', got %v", logEntry["level"])
-		}
-
-		if errVal, ok := logEntry["error"].(string); !ok || errVal != "test error" {
-			t.Errorf("expected error attribute 'test error', got %v", logEntry["error"])
-		}
 	})
 
-	t.Run("CollectsAndPlaysLogsWithContext", func(t *testing.T) {
-		// Create a buffer for capturing output
+	t.Run("ForwardToHandler", func(t *testing.T) {
 		var buf bytes.Buffer
-
-		// Create a JSON handler outputting to the buffer
 		jsonHandler := slog.NewJSONHandler(&buf, nil)
-
-		// Create collector with no output handler (nil)
-		collector := NewLogCollector(nil)
-
-		// Create logger that uses our collector
-		logger := slog.New(collector)
-
-		// Log some events
-		logger.WithGroup("app").Info("Starting up", "version", "1.0")
-		logger.Error("Something failed", "error", "test error")
-
-		// Now play the logs to the JSON handler with explicit context
-		ctx := context.Background()
-		if err := collector.PlayLogsCtx(ctx, jsonHandler); err != nil {
-			t.Fatalf("unexpected error playing logs with context: %v", err)
-		}
-
-		// Parse the output into JSON
-		output := buf.String()
-		lines := strings.Split(strings.TrimSpace(output), "\n")
-		if len(lines) != 2 {
-			t.Fatalf("expected 2 lines of JSON output, got %d", len(lines))
-		}
-
-		// Test the first log entry (info message)
-		var logEntry map[string]interface{}
-		if err := json.Unmarshal([]byte(lines[0]), &logEntry); err != nil {
-			t.Fatalf("invalid JSON: %v", err)
-		}
-
-		// Check for nested structure with groups
-		app, ok := logEntry["app"].(map[string]interface{})
-		if !ok {
-			t.Fatalf("Expected app group in output, got: %v", logEntry)
-		}
-
-		version, ok := app["version"].(string)
-		if !ok || version != "1.0" {
-			t.Errorf("Expected app.version=1.0, got: %v", app)
-		}
-	})
-
-	t.Run("ForwardsToUnderlyingHandler", func(t *testing.T) {
-		// Create a buffer for capturing immediate output
-		var forwardBuf bytes.Buffer
-		jsonHandler := slog.NewJSONHandler(&forwardBuf, nil)
-
-		// Create collector that forwards to jsonHandler
 		collector := NewLogCollector(jsonHandler)
 		logger := slog.New(collector)
 
-		// Log a message
 		logger.Info("This should be forwarded immediately")
 
-		// Verify the message was forwarded immediately
-		if forwardBuf.Len() == 0 {
+		if buf.Len() == 0 {
 			t.Error("expected output to be forwarded immediately")
 		}
 
-		var logEntry map[string]interface{}
-		if err := json.Unmarshal(forwardBuf.Bytes(), &logEntry); err != nil {
-			t.Fatalf("invalid JSON: %v", err)
-		}
-
-		if msg, ok := logEntry["msg"].(string); !ok || msg != "This should be forwarded immediately" {
-			t.Errorf("expected message was not forwarded, got %v", logEntry["msg"])
-		}
-
-		// Verify it was also collected
 		logs := collector.GetLogs()
 		if len(logs) != 1 {
 			t.Fatalf("expected 1 collected log, got %d", len(logs))
 		}
 	})
 
-	t.Run("WithAttrs", func(t *testing.T) {
-		collector := NewLogCollector(nil)
-
-		// Create a logger with simple attributes (no groups)
-		baseLogger := slog.New(collector)
-		attrLogger := baseLogger.With("requestID", "12345")
-
-		// Log with the attr logger
-		attrLogger.Info("Request processed")
-
-		// Get logs and verify
-		logs := collector.GetLogs()
-		if len(logs) != 1 {
-			t.Fatalf("expected 1 log, got %d", len(logs))
-		}
-
-		// Print the actual attributes for debugging
-		t.Logf("Log attributes: %v", logs[0].Attrs)
-
-		// Verify the log message was recorded correctly
-		if logs[0].Message != "Request processed" {
-			t.Errorf("expected message 'Request processed', got %q", logs[0].Message)
-		}
-
-		// Test log level
-		if logs[0].Level != slog.LevelInfo {
-			t.Errorf("expected level Info, got %v", logs[0].Level)
-		}
-	})
-
 	t.Run("EnabledLevels", func(t *testing.T) {
-		// Create a handler with INFO level
 		opts := &slog.HandlerOptions{Level: slog.LevelInfo}
 		var buf bytes.Buffer
 		jsonHandler := slog.NewJSONHandler(&buf, opts)
-
-		// Create collector
 		collector := NewLogCollector(jsonHandler)
 
-		// Verify that DEBUG is not enabled but INFO is
 		if collector.Enabled(context.Background(), slog.LevelDebug) {
 			t.Error("expected DEBUG level to be disabled")
 		}
-
 		if !collector.Enabled(context.Background(), slog.LevelInfo) {
 			t.Error("expected INFO level to be enabled")
 		}
 
-		// Collector with nil handler should enable all levels
 		nilCollector := NewLogCollector(nil)
 		if !nilCollector.Enabled(context.Background(), slog.LevelDebug) {
 			t.Error("expected all levels to be enabled with nil handler")
@@ -236,276 +95,212 @@ func TestLogCollector(t *testing.T) {
 	})
 }
 
-// errorHandler is a test handler that returns an error on Handle
-type errorHandler struct{}
+// Attributes and groups - the critical test that was flaky
+func TestAttributesAndGroups(t *testing.T) {
+	t.Run("WithGroupAndAttributes", func(t *testing.T) {
+		var origBuf, replayBuf bytes.Buffer
 
-func (h *errorHandler) Enabled(ctx context.Context, level slog.Level) bool {
-	return true
-}
-
-func (h *errorHandler) Handle(ctx context.Context, r slog.Record) error {
-	return io.ErrUnexpectedEOF // Return an error for testing
-}
-
-func (h *errorHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
-	return h
-}
-
-func (h *errorHandler) WithGroup(name string) slog.Handler {
-	return h
-}
-
-// sleepyHandler is a test handler that sleeps before handling records
-type sleepyHandler struct {
-	sleepTime time.Duration
-}
-
-func (h *sleepyHandler) Enabled(ctx context.Context, level slog.Level) bool {
-	return true
-}
-
-func (h *sleepyHandler) Handle(ctx context.Context, r slog.Record) error {
-	time.Sleep(h.sleepTime)
-	// Check if context was canceled during sleep
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	default:
-		return nil
-	}
-}
-
-func (h *sleepyHandler) WithAttrs(attrs []slog.Attr) slog.Handler {
-	return h
-}
-
-func (h *sleepyHandler) WithGroup(name string) slog.Handler {
-	return h
-}
-
-func TestStorageTypes(t *testing.T) {
-	// Verify we can work with the storage types
-	var record storage.Record
-	record.Message = "test"
-	record.Level = slog.LevelInfo
-
-	if record.Message != "test" || record.Level != slog.LevelInfo {
-		t.Error("Failed to use storage.Record type")
-	}
-}
-
-func TestStorageCapacity(t *testing.T) {
-	t.Run("InitialCapacity", func(t *testing.T) {
-		// Create a collector
-		collector := NewLogCollector(nil)
-
-		// Verify no logs initially
-		logs := collector.GetLogs()
-		if len(logs) != 0 {
-			t.Errorf("Expected initial log count to be 0, got %d", len(logs))
-		}
-	})
-
-	t.Run("GrowsAsNeeded", func(t *testing.T) {
-		// Create a collector
-		collector := NewLogCollector(nil)
+		textHandler := slog.NewTextHandler(&origBuf, nil)
+		collector := NewLogCollector(textHandler)
 		logger := slog.New(collector)
 
-		// Add more logs than the default capacity
-		logsToAdd := 20 // Default capacity is 10, so this is 2x default
-		for i := 0; i < logsToAdd; i++ {
-			logger.Info("test log", "index", i)
+		baseLogger := logger.With("global", "value")
+		groupLogger := baseLogger.WithGroup("group1")
+		groupWithAttrLogger := groupLogger.With("attribute", "value")
+		nestedGroupLogger := logger.WithGroup("parent").WithGroup("child")
+		multiAttrLogger := logger.WithGroup("multi").With("attr1", "value1", "attr2", "value2")
+
+		baseLogger.Info("Base log")
+		groupLogger.Info("Group log", "field1", "value1")
+		groupWithAttrLogger.Error("Group with attr log", "field2", "value2")
+		nestedGroupLogger.Warn("Nested group log", "nested", "value")
+		multiAttrLogger.Info("Multiple attrs", "extra", "value")
+
+		origOutput := origBuf.String()
+		origLogs := parseLogOutput(t, origOutput)
+
+		replayHandler := slog.NewTextHandler(&replayBuf, nil)
+		if err := collector.PlayLogs(replayHandler); err != nil {
+			t.Fatalf("Failed to replay logs: %v", err)
 		}
 
-		// Verify all logs were stored
+		replayOutput := replayBuf.String()
+		replayLogs := parseLogOutput(t, replayOutput)
+
+		if len(origLogs) != len(replayLogs) {
+			t.Errorf("Different number of log lines: original=%d, replayed=%d",
+				len(origLogs), len(replayLogs))
+		}
+
+		for timestamp, origFields := range origLogs {
+			replayFields, found := replayLogs[timestamp]
+			if !found {
+				t.Errorf("Log entry with timestamp %s missing in replay", timestamp)
+				continue
+			}
+			compareLogFields(t, timestamp, origFields, replayFields)
+		}
+	})
+
+	t.Run("GetLogsReturnsCompleteRecords", func(t *testing.T) {
+		collector := NewLogCollector(nil)
+
+		logger := slog.New(collector.WithAttrs([]slog.Attr{
+			slog.String("service", "api"),
+			slog.Int("version", 2),
+		}))
+
+		logger.Info("request handled", "duration", 100, "status", 200)
+
 		logs := collector.GetLogs()
-		if len(logs) != logsToAdd {
-			t.Errorf("Expected %d logs, got %d", logsToAdd, len(logs))
+		if len(logs) != 1 {
+			t.Fatalf("Expected 1 log, got %d", len(logs))
+		}
+
+		attrs := make(map[string]any)
+		for _, attr := range logs[0].Attrs {
+			attrs[attr.Key] = attr.Value.Any()
+		}
+
+		if v, ok := attrs["service"]; !ok || v != "api" {
+			t.Errorf("Missing or incorrect service attribute: %v", v)
+		}
+		if v, ok := attrs["version"]; !ok || v != int64(2) {
+			t.Errorf("Missing or incorrect version attribute: %v", v)
+		}
+		if v, ok := attrs["duration"]; !ok || v != int64(100) {
+			t.Errorf("Missing or incorrect duration attribute: %v", v)
+		}
+		if v, ok := attrs["status"]; !ok || v != int64(200) {
+			t.Errorf("Missing or incorrect status attribute: %v", v)
+		}
+	})
+
+	t.Run("GroupsProperlyApplied", func(t *testing.T) {
+		collector := NewLogCollector(nil)
+
+		logger := slog.New(
+			collector.
+				WithAttrs([]slog.Attr{slog.String("global", "value")}).
+				WithGroup("server").
+				WithAttrs([]slog.Attr{slog.String("host", "localhost")}),
+		)
+
+		logger.Info("server started", "port", 8080)
+
+		logs := collector.GetLogs()
+		if len(logs) != 1 {
+			t.Fatalf("Expected 1 log, got %d", len(logs))
+		}
+
+		attrs := make(map[string]any)
+		for _, attr := range logs[0].Attrs {
+			flattenAttrs(attr, "", attrs)
+		}
+
+		if v, ok := attrs["global"]; !ok || v != "value" {
+			t.Errorf("Missing or incorrect global attribute: %v", v)
+		}
+		if v, ok := attrs["server.host"]; !ok || v != "localhost" {
+			t.Errorf("Missing or incorrect server.host attribute: %v", v)
+		}
+		if v, ok := attrs["server.port"]; !ok || v != int64(8080) {
+			t.Errorf("Missing or incorrect server.port attribute: %v", v)
 		}
 	})
 }
 
-func TestEdgeCases(t *testing.T) {
+// Error handling and edge cases
+func TestErrorHandling(t *testing.T) {
 	t.Run("WithAttrsEmptySlice", func(t *testing.T) {
-		// Test with empty attrs slice
 		collector := NewLogCollector(nil)
 		newHandler := collector.WithAttrs([]slog.Attr{})
-
-		// Should return the same handler instance
 		if newHandler != collector {
 			t.Error("Expected WithAttrs with empty slice to return same handler")
 		}
 	})
 
 	t.Run("WithGroupEmptyName", func(t *testing.T) {
-		// Test with empty group name
 		collector := NewLogCollector(nil)
 		newHandler := collector.WithGroup("")
-
-		// Should return the same handler instance
 		if newHandler != collector {
 			t.Error("Expected WithGroup with empty name to return same handler")
 		}
 	})
 
-	t.Run("PlayLogsNilStore", func(t *testing.T) {
-		// Create collector with nil store (unusual, but possible in edge cases)
-		collector := &LogCollector{
-			store:   nil,
-			handler: nil,
-		}
-
-		// Should return nil without panicking
-		err := collector.PlayLogs(slog.NewTextHandler(io.Discard, nil))
-		if err != nil {
-			t.Errorf("Expected PlayLogs with nil store to return nil, got %v", err)
-		}
-	})
-
-	t.Run("PlayLogsEmptyStore", func(t *testing.T) {
-		// Create a collector with a valid but empty store
-		collector := NewLogCollector(nil)
-
-		// Should return nil without panicking
-		err := collector.PlayLogs(slog.NewTextHandler(io.Discard, nil))
-		if err != nil {
-			t.Errorf("Expected PlayLogs with empty store to return nil, got %v", err)
-		}
-	})
-
 	t.Run("PlayLogsNilHandler", func(t *testing.T) {
-		// Create a collector with some data
 		collector := NewLogCollector(nil)
 		logger := slog.New(collector)
 		logger.Info("Test message")
 
-		// Should return error for nil handler
 		err := collector.PlayLogs(nil)
 		if err == nil || !strings.Contains(err.Error(), "handler is nil") {
 			t.Errorf("Expected error for nil handler, got: %v", err)
 		}
 	})
 
-	t.Run("PlayLogsCtxNilHandler", func(t *testing.T) {
-		// Create a collector with some data
-		collector := NewLogCollector(nil)
-		logger := slog.New(collector)
-		logger.Info("Test message")
-
-		// Should return error for nil handler
-		err := collector.PlayLogsCtx(context.Background(), nil)
-		if err == nil || !strings.Contains(err.Error(), "handler is nil") {
-			t.Errorf("Expected error for nil handler, got: %v", err)
-		}
-	})
-
 	t.Run("PlayLogsErrorHandler", func(t *testing.T) {
-		// Create a collector with some test data
 		collector := NewLogCollector(nil)
 		logger := slog.New(collector)
 		logger.Info("Test message")
 
-		// Verify we have logs to replay
-		if len(collector.GetLogs()) == 0 {
-			t.Fatal("Expected collector to have logs")
-		}
-
-		// Attempt to play logs with a handler that returns errors
 		errHandler := &errorHandler{}
 		err := collector.PlayLogs(errHandler)
 
-		// Should propagate the error
 		if err != io.ErrUnexpectedEOF {
 			t.Errorf("Expected PlayLogs to return io.ErrUnexpectedEOF, got %v", err)
 		}
 	})
 
-	t.Run("PlayLogsCtxCancellation", func(t *testing.T) {
-		// Create a collector with multiple logs
+	t.Run("ContextCancellation", func(t *testing.T) {
 		collector := NewLogCollector(nil)
 		logger := slog.New(collector)
 
-		// Add several logs to ensure we have enough to process
 		for i := 0; i < 5; i++ {
 			logger.Info("Test message", "index", i)
 		}
 
-		// Create a canceled context
 		ctx, cancel := context.WithCancel(context.Background())
-		cancel() // Cancel immediately
+		cancel()
 
-		// Should return context.Canceled error
 		err := collector.PlayLogsCtx(ctx, slog.NewTextHandler(io.Discard, nil))
 		if err != context.Canceled {
 			t.Errorf("Expected context.Canceled error, got: %v", err)
 		}
 	})
 
-	t.Run("PlayLogsCtxTimeout", func(t *testing.T) {
-		// Create a collector with some test data
+	t.Run("ContextTimeout", func(t *testing.T) {
 		collector := NewLogCollector(nil)
 		logger := slog.New(collector)
 		logger.Info("Test message")
 
-		// Create a context with timeout
 		ctx, cancel := context.WithTimeout(context.Background(), 50*time.Millisecond)
 		defer cancel()
 
-		// Create a handler that sleeps to trigger timeout
-		slowHandler := &sleepyHandler{
-			sleepTime: 100 * time.Millisecond,
-		}
-
-		// Should return context.DeadlineExceeded error
+		slowHandler := &sleepyHandler{sleepTime: 100 * time.Millisecond}
 		err := collector.PlayLogsCtx(ctx, slowHandler)
 		if err != context.DeadlineExceeded {
 			t.Errorf("Expected context.DeadlineExceeded error, got: %v", err)
 		}
 	})
 
-	t.Run("GetLogsNilStore", func(t *testing.T) {
-		// Create collector with nil store
-		collector := &LogCollector{
-			store:   nil,
-			handler: nil,
-		}
-
-		// Should return nil
-		logs := collector.GetLogs()
-		if logs != nil {
-			t.Errorf("Expected GetLogs with nil store to return nil, got %v", logs)
-		}
-	})
-
 	t.Run("GetLogsCopiesData", func(t *testing.T) {
-		// Create a collector with some data
 		collector := NewLogCollector(nil)
 		logger := slog.New(collector)
 		logger.Info("test message")
 
-		// Get logs
 		logs1 := collector.GetLogs()
 		if len(logs1) != 1 {
 			t.Fatalf("Expected 1 log, got %d", len(logs1))
 		}
 
-		// Get logs again - should be a separate copy
 		logs2 := collector.GetLogs()
-
-		// Add another log
 		logger.Info("second message")
 
-		// Original logs1 should be unaffected
-		if len(logs1) != 1 {
-			t.Errorf("Expected first GetLogs result to be unchanged, got %d logs", len(logs1))
+		if len(logs1) != 1 || len(logs2) != 1 {
+			t.Error("Expected previous GetLogs results to be unchanged")
 		}
 
-		// logs2 should also be unaffected
-		if len(logs2) != 1 {
-			t.Errorf("Expected second GetLogs result to be unchanged, got %d logs", len(logs2))
-		}
-
-		// New call to GetLogs should have all logs
 		logs3 := collector.GetLogs()
 		if len(logs3) != 2 {
 			t.Errorf("Expected final GetLogs to have 2 logs, got %d", len(logs3))
@@ -513,101 +308,283 @@ func TestEdgeCases(t *testing.T) {
 	})
 }
 
-func TestComplexScenarios(t *testing.T) {
-	t.Run("NestedGroups", func(t *testing.T) {
-		// Create a buffer to capture output
-		var buf bytes.Buffer
-		jsonHandler := slog.NewJSONHandler(&buf, nil)
-
-		// Create collector with nil handler (just for collection)
+// Concurrency tests
+func TestConcurrency(t *testing.T) {
+	t.Run("ConcurrentLoggingWithSharedStorage", func(t *testing.T) {
 		collector := NewLogCollector(nil)
 
-		// Create a logger with nested groups
-		baseLogger := slog.New(collector)
-		parentGroupLogger := baseLogger.WithGroup("parent")
-		nestedGroupLogger := parentGroupLogger.WithGroup("child")
+		collector1 := collector.WithAttrs([]slog.Attr{slog.String("collector", "1")})
+		collector2 := collector.WithAttrs([]slog.Attr{slog.String("collector", "2")})
+		collector3 := collector.WithGroup("group1")
 
-		// Log with the nested logger
-		nestedGroupLogger.Info("nested log", "key", "value")
+		logger1 := slog.New(collector1)
+		logger2 := slog.New(collector2)
+		logger3 := slog.New(collector3)
 
-		// Now play logs to JSON handler
-		buf.Reset()
-		err := collector.PlayLogs(jsonHandler)
-		if err != nil {
-			t.Fatalf("unexpected error playing logs: %v", err)
-		}
+		var wg sync.WaitGroup
+		iterations := 100
+		wg.Add(3)
 
-		// Parse the JSON output
-		var result map[string]interface{}
-		if err := json.Unmarshal(buf.Bytes(), &result); err != nil {
-			t.Fatalf("Failed to parse JSON: %v", err)
-		}
+		go func() {
+			defer wg.Done()
+			for i := range iterations {
+				logger1.Info("test message", "index", i, "logger", 1)
+			}
+		}()
 
-		// Check for nested structure: parent.child.key
-		parent, ok := result["parent"].(map[string]interface{})
-		if !ok {
-			t.Fatalf("Expected parent group in output, got: %v", result)
-		}
+		go func() {
+			defer wg.Done()
+			for i := range iterations {
+				logger2.Info("test message", "index", i, "logger", 2)
+			}
+		}()
 
-		child, ok := parent["child"].(map[string]interface{})
-		if !ok {
-			t.Fatalf("Expected child group in output, got: %v", parent)
-		}
+		go func() {
+			defer wg.Done()
+			for i := range iterations {
+				logger3.Info("test message", "index", i, "logger", 3)
+			}
+		}()
 
-		value, ok := child["key"].(string)
-		if !ok || value != "value" {
-			t.Errorf("Expected child.key=value, got: %v", child)
+		wg.Wait()
+
+		logs := collector.GetLogs()
+		expectedCount := iterations * 3
+		if len(logs) != expectedCount {
+			t.Errorf("Expected %d logs, got %d", expectedCount, len(logs))
 		}
 	})
 
-	t.Run("MultipleAttributes", func(t *testing.T) {
-		// Create collector
+	t.Run("ConcurrentPlayLogs", func(t *testing.T) {
 		collector := NewLogCollector(nil)
-
-		// Create a logger with attributes
 		logger := slog.New(collector)
-		attrLogger := logger.With(
-			"string", "value",
-			"int", 42,
-			"bool", true,
-			"timestamp", time.Date(2025, 5, 14, 12, 0, 0, 0, time.UTC),
-		)
 
-		// Log with the attribute-rich logger
-		attrLogger.Info("attribute test")
+		for i := range 100 {
+			logger.Info("test message", "index", i)
+		}
 
-		// Get logs and verify all attributes were stored
+		var wg sync.WaitGroup
+		wg.Add(3)
+
+		for range 3 {
+			go func() {
+				defer wg.Done()
+				handler := slog.NewTextHandler(&discardWriter{}, nil)
+				err := collector.PlayLogs(handler)
+				if err != nil {
+					t.Errorf("PlayLogs failed: %v", err)
+				}
+			}()
+		}
+
+		wg.Wait()
+	})
+}
+
+// PC preservation tests
+func TestPCPreservation(t *testing.T) {
+	t.Run("BasicPCPreservation", func(t *testing.T) {
+		collector := NewLogCollector(nil)
+		logger := slog.New(collector)
+
+		logger.Info("test message with PC")
+
 		logs := collector.GetLogs()
 		if len(logs) != 1 {
-			t.Fatalf("expected 1 log, got %d", len(logs))
+			t.Fatalf("Expected 1 log, got %d", len(logs))
 		}
 
-		// Log the attributes we received to help with debugging
-		t.Logf("Got %d attributes:", len(logs[0].Attrs))
-		for i, attr := range logs[0].Attrs {
-			t.Logf("  Attr[%d]: %s = %v", i, attr.Key, attr.Value)
+		if logs[0].PC == 0 {
+			t.Error("Expected non-zero PC value")
 		}
 
-		// The slog.With method may attach attributes differently than we expect
-		// in different Go versions. For now, just verify the message is logged.
-		if logs[0].Message != "attribute test" {
-			t.Errorf("Expected message 'attribute test', got %q", logs[0].Message)
+		var buf bytes.Buffer
+		handler := slog.NewTextHandler(&buf, &slog.HandlerOptions{AddSource: true})
+
+		err := collector.PlayLogs(handler)
+		if err != nil {
+			t.Fatalf("PlayLogs failed: %v", err)
 		}
 
-		// Verify it's the right level
-		if logs[0].Level != slog.LevelInfo {
-			t.Errorf("Expected INFO level, got %v", logs[0].Level)
+		output := buf.String()
+		if !strings.Contains(output, "loglater_test.go") {
+			t.Errorf("Expected source file in output, got: %s", output)
+		}
+	})
+
+	t.Run("PCWithGroupsAndAttrs", func(t *testing.T) {
+		collector := NewLogCollector(nil)
+
+		logger := slog.New(collector.WithGroup("testgroup").WithAttrs([]slog.Attr{
+			slog.String("component", "test"),
+		}))
+
+		logger.Info("grouped message with attrs")
+
+		logs := collector.GetLogs()
+		if len(logs) != 1 {
+			t.Fatalf("Expected 1 log, got %d", len(logs))
+		}
+
+		if logs[0].PC == 0 {
+			t.Error("Expected non-zero PC value with groups and attrs")
 		}
 	})
 }
 
-// parseLogLine breaks a log line into fields
+// Storage behavior and consistency tests
+func TestStorageBehavior(t *testing.T) {
+	t.Run("StorageReturnsRawRecords", func(t *testing.T) {
+		store := storage.NewRecordStorage()
+
+		record := storage.Record{
+			Time:    time.Now(),
+			Level:   slog.LevelInfo,
+			Message: "test",
+			Attrs:   []slog.Attr{slog.String("msg_attr", "value")},
+			Sequence: storage.HandlerSequence{
+				{Type: "attrs", Attrs: []slog.Attr{slog.String("global", "value")}},
+				{Type: "group", Group: "g1"},
+				{Type: "attrs", Attrs: []slog.Attr{slog.String("grouped", "value")}},
+			},
+		}
+
+		store.Append(&record)
+
+		records := store.GetAll()
+		if len(records) != 1 {
+			t.Fatalf("Expected 1 record, got %d", len(records))
+		}
+
+		if len(records[0].Attrs) != 1 {
+			t.Errorf("Expected 1 attribute in raw record, got %d", len(records[0].Attrs))
+		}
+
+		if records[0].Attrs[0].Key != "msg_attr" {
+			t.Errorf("Expected msg_attr in raw record, got %s", records[0].Attrs[0].Key)
+		}
+
+		if len(records[0].Sequence) != 3 {
+			t.Errorf("Expected sequence of 3 operations, got %d", len(records[0].Sequence))
+		}
+	})
+
+	t.Run("RecordRealize", func(t *testing.T) {
+		record := storage.Record{
+			Time:    time.Now(),
+			Level:   slog.LevelInfo,
+			Message: "test",
+			Attrs:   []slog.Attr{slog.String("msg", "value")},
+			Sequence: storage.HandlerSequence{
+				{Type: "attrs", Attrs: []slog.Attr{slog.String("global", "value")}},
+				{Type: "group", Group: "g1"},
+				{Type: "attrs", Attrs: []slog.Attr{slog.String("grouped", "value")}},
+			},
+		}
+
+		realized := record.Realize()
+
+		attrs := make(map[string]any)
+		for _, attr := range realized.Attrs {
+			flattenAttrs(attr, "", attrs)
+		}
+
+		if v, ok := attrs["global"]; !ok || v != "value" {
+			t.Errorf("Missing or incorrect global attribute: %v", v)
+		}
+		if v, ok := attrs["g1.grouped"]; !ok || v != "value" {
+			t.Errorf("Missing or incorrect g1.grouped attribute: %v", v)
+		}
+		if v, ok := attrs["g1.msg"]; !ok || v != "value" {
+			t.Errorf("Missing or incorrect g1.msg attribute: %v", v)
+		}
+	})
+
+	t.Run("GetLogsPlayLogsConsistency", func(t *testing.T) {
+		collector := NewLogCollector(nil)
+
+		logger := slog.New(
+			collector.
+				WithAttrs([]slog.Attr{slog.String("service", "api")}).
+				WithGroup("request").
+				WithAttrs([]slog.Attr{slog.String("id", "123")}),
+		)
+
+		logger.Info("start", "method", "GET")
+		logger.WithGroup("response").Info("complete", "status", 200)
+
+		logs := collector.GetLogs()
+		if len(logs) != 2 {
+			t.Fatalf("Expected 2 logs, got %d", len(logs))
+		}
+
+		attrs1 := make(map[string]any)
+		for _, attr := range logs[0].Attrs {
+			flattenAttrs(attr, "", attrs1)
+		}
+
+		if v, ok := attrs1["service"]; !ok || v != "api" {
+			t.Errorf("First log missing service=api: %v", attrs1)
+		}
+		if v, ok := attrs1["request.id"]; !ok || v != "123" {
+			t.Errorf("First log missing request.id=123: %v", attrs1)
+		}
+		if v, ok := attrs1["request.method"]; !ok || v != "GET" {
+			t.Errorf("First log missing request.method=GET: %v", attrs1)
+		}
+
+		var buf bytes.Buffer
+		textHandler := slog.NewTextHandler(&buf, &slog.HandlerOptions{
+			ReplaceAttr: func(groups []string, a slog.Attr) slog.Attr {
+				if a.Key == "time" {
+					return slog.Attr{}
+				}
+				return a
+			},
+		})
+
+		if err := collector.PlayLogs(textHandler); err != nil {
+			t.Fatalf("PlayLogs failed: %v", err)
+		}
+
+		output := buf.String()
+		lines := strings.Split(strings.TrimSpace(output), "\n")
+		if len(lines) != 2 {
+			t.Fatalf("Expected 2 log lines, got %d: %s", len(lines), output)
+		}
+
+		line1 := lines[0]
+		if !strings.Contains(line1, "service=api") {
+			t.Errorf("First line missing service=api: %s", line1)
+		}
+		if !strings.Contains(line1, "request.id=123") {
+			t.Errorf("First line missing request.id=123: %s", line1)
+		}
+		if !strings.Contains(line1, "request.method=GET") {
+			t.Errorf("First line missing request.method=GET: %s", line1)
+		}
+	})
+}
+
+// Helper functions
+func flattenAttrs(attr slog.Attr, prefix string, result map[string]any) {
+	key := attr.Key
+	if prefix != "" {
+		key = prefix + "." + key
+	}
+
+	if attr.Value.Kind() == slog.KindGroup {
+		for _, groupAttr := range attr.Value.Group() {
+			flattenAttrs(groupAttr, key, result)
+		}
+	} else {
+		result[key] = attr.Value.Any()
+	}
+}
+
 func parseLogLine(t *testing.T, line string) map[string]string {
 	t.Helper()
-
 	result := make(map[string]string)
-
-	// Split by space, but respect quoted values
 	var fields []string
 	var currentField strings.Builder
 	inQuote := false
@@ -626,12 +603,10 @@ func parseLogLine(t *testing.T, line string) map[string]string {
 		}
 	}
 
-	// Add the last field if any
 	if currentField.Len() > 0 {
 		fields = append(fields, currentField.String())
 	}
 
-	// Parse each field into key=value pairs
 	for _, field := range fields {
 		if idx := strings.Index(field, "="); idx >= 0 {
 			key := field[:idx]
@@ -643,10 +618,8 @@ func parseLogLine(t *testing.T, line string) map[string]string {
 	return result
 }
 
-// parseLogOutput parses multiple log lines into a map using timestamp as key
 func parseLogOutput(t *testing.T, output string) map[string]map[string]string {
 	t.Helper()
-
 	lines := strings.Split(strings.TrimSpace(output), "\n")
 	result := make(map[string]map[string]string)
 
@@ -660,56 +633,9 @@ func parseLogOutput(t *testing.T, output string) map[string]map[string]string {
 	return result
 }
 
-// compareLogFields compares fields between original and replayed logs
 func compareLogFields(t *testing.T, timestamp string, origFields, replayFields map[string]string) {
 	t.Helper()
 
-	// Compare fields in each log entry
-	// First check for fields that start with a group prefix, which is our focus
-	var origGroupFields, replayGroupFields []string
-
-	for field := range origFields {
-		// Look for fields that would be prefixed with group names
-		parts := strings.Split(field, ".")
-		if len(parts) > 1 {
-			origGroupFields = append(origGroupFields, field)
-		}
-	}
-
-	for field := range replayFields {
-		parts := strings.Split(field, ".")
-		if len(parts) > 1 {
-			replayGroupFields = append(replayGroupFields, field)
-		}
-	}
-
-	// Sort for reliable comparison
-	sort.Strings(origGroupFields)
-	sort.Strings(replayGroupFields)
-
-	// Check for missing group fields in replay
-	for _, field := range origGroupFields {
-		if _, exists := replayFields[field]; !exists {
-			t.Errorf("Field %q present in original but missing in replay for timestamp %s",
-				field, timestamp)
-		}
-	}
-
-	// Report different number of group fields
-	if len(origGroupFields) != len(replayGroupFields) {
-		t.Errorf("Different number of group fields for timestamp %s: original=%d, replayed=%d",
-			timestamp, len(origGroupFields), len(replayGroupFields))
-
-		// Show the differences
-		t.Logf("Original group fields: %v", origGroupFields)
-		t.Logf("Replayed group fields: %v", replayGroupFields)
-
-		// Add more detailed logging to debug non-determinism
-		t.Logf("All original fields: %v", origFields)
-		t.Logf("All replayed fields: %v", replayFields)
-	}
-
-	// Also compare all field values
 	for field, origValue := range origFields {
 		if replayValue, ok := replayFields[field]; ok {
 			if origValue != replayValue {
@@ -720,121 +646,33 @@ func compareLogFields(t *testing.T, timestamp string, origFields, replayFields m
 	}
 }
 
-func TestGroupPreservation(t *testing.T) {
-	// Create a buffer to capture output
-	var buf bytes.Buffer
-	jsonHandler := slog.NewJSONHandler(&buf, &slog.HandlerOptions{
-		ReplaceAttr: func(groups []string, a slog.Attr) slog.Attr {
-			// This callback helps inspect the structure
-			return a
-		},
-	})
+// Test helper types
+type errorHandler struct{}
 
-	// Create collector and base logger
-	collector := NewLogCollector(nil)
-	baseLogger := slog.New(collector)
+func (h *errorHandler) Enabled(ctx context.Context, level slog.Level) bool { return true }
+func (h *errorHandler) Handle(ctx context.Context, r slog.Record) error    { return io.ErrUnexpectedEOF }
+func (h *errorHandler) WithAttrs(attrs []slog.Attr) slog.Handler           { return h }
+func (h *errorHandler) WithGroup(name string) slog.Handler                 { return h }
 
-	// Create a logger with a group
-	groupLogger := baseLogger.WithGroup("testgroup")
-
-	// Log with the group logger
-	groupLogger.Info("test message", "key", "value")
-
-	// Verify the log was captured
-	logs := collector.GetLogs()
-	if len(logs) != 1 {
-		t.Fatalf("expected 1 log, got %d", len(logs))
-	}
-
-	// Log all captured attributes to understand what we have
-	t.Logf("Log message: %s", logs[0].Message)
-	t.Logf("Log level: %s", logs[0].Level)
-	t.Logf("Number of attributes: %d", len(logs[0].Attrs))
-	for i, attr := range logs[0].Attrs {
-		t.Logf("Attr[%d]: key=%q value=%v", i, attr.Key, attr.Value)
-	}
-
-	// Now test replaying with group structure preserved
-	buf.Reset()
-	err := collector.PlayLogs(jsonHandler)
-	if err != nil {
-		t.Fatalf("error playing logs: %v", err)
-	}
-
-	// Check the output
-	output := buf.String()
-	t.Logf("JSON output: %s", output)
-
-	// Parse the JSON output
-	var result map[string]interface{}
-	if err := json.Unmarshal([]byte(output), &result); err != nil {
-		t.Fatalf("Failed to parse JSON: %v", err)
-	}
-
-	// This test should fail initially since we haven't implemented group handling yet
-	if _, ok := result["testgroup"]; !ok {
-		t.Errorf("Expected 'testgroup' in output JSON, but it wasn't found")
-	}
+type sleepyHandler struct {
+	sleepTime time.Duration
 }
 
-func TestWithGroupAndAttributes(t *testing.T) {
-	// Create buffers to capture output
-	var origBuf, replayBuf bytes.Buffer
-
-	// Create a collector with a text handler for direct output
-	textHandler := slog.NewTextHandler(&origBuf, nil)
-	collector := NewLogCollector(textHandler)
-	logger := slog.New(collector)
-
-	// Create different types of loggers with groups and attributes
-	baseLogger := logger.With("global", "value")
-	groupLogger := baseLogger.WithGroup("group1")
-	groupWithAttrLogger := groupLogger.With("attribute", "value")
-	nestedGroupLogger := logger.WithGroup("parent").WithGroup("child")
-	multiAttrLogger := logger.WithGroup("multi").With("attr1", "value1", "attr2", "value2")
-
-	// Log messages with different loggers
-	baseLogger.Info("Base log")
-	groupLogger.Info("Group log", "field1", "value1")
-	groupWithAttrLogger.Error("Group with attr log", "field2", "value2")
-	nestedGroupLogger.Warn("Nested group log", "nested", "value")
-	multiAttrLogger.Info("Multiple attrs", "extra", "value")
-
-	// Get all original output lines
-	origOutput := origBuf.String()
-	t.Logf("Original output:\n%s", origOutput)
-
-	// Parse original output
-	origLogs := parseLogOutput(t, origOutput)
-
-	// Now replay the logs to a new text handler
-	replayHandler := slog.NewTextHandler(&replayBuf, nil)
-	if err := collector.PlayLogs(replayHandler); err != nil {
-		t.Fatalf("Failed to replay logs: %v", err)
+func (h *sleepyHandler) Enabled(ctx context.Context, level slog.Level) bool { return true }
+func (h *sleepyHandler) Handle(ctx context.Context, r slog.Record) error {
+	time.Sleep(h.sleepTime)
+	select {
+	case <-ctx.Done():
+		return ctx.Err()
+	default:
+		return nil
 	}
+}
+func (h *sleepyHandler) WithAttrs(attrs []slog.Attr) slog.Handler { return h }
+func (h *sleepyHandler) WithGroup(name string) slog.Handler       { return h }
 
-	// Get all replayed output lines
-	replayOutput := replayBuf.String()
-	t.Logf("Replayed output:\n%s", replayOutput)
+type discardWriter struct{}
 
-	// Parse replayed output
-	replayLogs := parseLogOutput(t, replayOutput)
-
-	// Verify we have the same number of log lines
-	if len(origLogs) != len(replayLogs) {
-		t.Errorf("Different number of log lines: original=%d, replayed=%d",
-			len(origLogs), len(replayLogs))
-	}
-
-	// Compare each log entry by timestamp
-	for timestamp, origFields := range origLogs {
-		replayFields, found := replayLogs[timestamp]
-		if !found {
-			t.Errorf("Log entry with timestamp %s missing in replay", timestamp)
-			continue
-		}
-
-		// Compare the fields for this log entry
-		compareLogFields(t, timestamp, origFields, replayFields)
-	}
+func (d *discardWriter) Write(p []byte) (n int, err error) {
+	return len(p), nil
 }
